+++ openedx_webhooks/__init__.py
@@ -2,6 +2,6 @@
 Webhooks for Open edX.
 """
 
-__version__ = '1.0.2'
+__version__ = '2.0.0'
 
 default_app_config = 'openedx_webhooks.apps.WebhooksConfig'  # pylint: disable=invalid-name