+++ openedx_webhooks/settings/common.py
@@ -7,7 +7,7 @@ Common Pluggable Django App settings.
 
 def plugin_settings(settings):
     """
-    Inject local settings into django settings.
+    Declare all filters and their handlers.
     """
 
     settings.OPEN_EDX_FILTERS_CONFIG = {
@@ -71,4 +71,84 @@ def plugin_settings(settings):
                 "openedx_webhooks.filters.DashboardRenderStartedWebFilter"
             ]
         },
+
+        "org.openedx.learning.vertical_block_child.render.started.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.VerticalBlockChildRenderStartedWebFilter"
+            ]
+        },
+        "org.openedx.learning.course_enrollment_queryset.requested.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.CourseEnrollmentQuerysetRequestedWebFilter"
+            ]
+        },
+        "org.openedx.learning.xblock.render.started.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.RenderXBlockStartedWebFilter"
+            ]
+        },
+        "org.openedx.learning.vertical_block.render.completed.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.VerticalBlockRenderCompletedWebFilter"
+            ]
+        },
+        "org.openedx.learning.course.homepage.url.creation.started.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.CourseHomeUrlCreationStartedWebFilter"
+            ]
+        },
+        "org.openedx.learning.home.enrollment.api.rendered.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.CourseEnrollmentAPIRenderStartedWebFilter"
+            ]
+        },
+        "org.openedx.learning.home.courserun.api.rendered.started.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.CourseRunAPIRenderStartedWebFilter"
+            ]
+        },
+        "org.openedx.learning.instructor.dashboard.render.started.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.InstructorDashboardRenderStartedWebFilter"
+            ]
+        },
+        "org.openedx.learning.ora.submission_view.render.started.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.ORASubmissionViewRenderStartedWebFilter"
+            ]
+        },
+        "org.openedx.learning.idv.page.url.requested.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.IDVPageURLRequestedWebFilter"
+            ]
+        },
+        "org.openedx.learning.course_about.page.url.requested.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.CourseAboutPageURLRequestedWebFilter"
+            ]
+        },
+        "org.openedx.learning.schedule.queryset.requested.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.ScheduleQuerySetRequestedWebFilter"
+            ]
+        },
+        "org.openedx.content_authoring.lms.page.url.requested.v1": {
+            "fail_silently": False,
+            "pipeline": [
+                "openedx_webhooks.filters.LMSPageURLRequestedWebFilter"
+            ]
+        },
+
     }